feat: create kneatr database
This commit introduces the `KneatrDatabase` class, an abstract class that extends `RoomDatabase`. It defines the entities for the database, including `ContactEntity`, `ContactTierEntity`, `TagEntity`, and `ContactTagCrossRef`. The database version is set to 1, and schema export is disabled.

The `KneatrDatabase` class also includes abstract functions for accessing the DAOs: `contactDao()`, `contactTierDao()`, `tagDao()`, and `contactTagCrossRefDao()`. Additionally, it utilizes `LocalDateConverters` for type conversion.

Signed-off-by: Tobi Olutimehin <olutimehintobi@gmail.com>