feat: Add contact extensions and date time helper
This commit introduces `ContactExtensions` for calculating `lastContactedDate`, `nextContactDate`, and `isOverdue` for contacts. It also adds a `DateTimeHelper` for date calculations and retrieving the current date.

These changes allow for more robust date-based logic related to contacts, such as determining when a contact is overdue for communication.

Corresponding tests for both `ContactExtensions` and `DateTimeHelper` have been added.

The `ContactEntity` has been updated to mark `lastContactedDate` and `nextContactDate` as fields to be removed soon, as this logic will be handled by the new extensions.

The `ContactWithTagsAndTier` relation has been updated to include `communicationLogs`.

The test for querying overdue contacts in `ContactDaoTest` has been commented out, likely to be updated or replaced with logic using the new extensions.

Signed-off-by: Tobi Olutimehin <olutimehintobi@gmail.com>